# docs/openapi/flowcept-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Flowcept Webservice API",
    "description": "Read-only REST API for Flowcept provenance data. Provides workflows, tasks, and objects endpoints with query support.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/datasets": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "List Datasets",
        "description": "List dataset objects with optional filters.",
        "operationId": "list_datasets_api_v1_datasets_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "workflow_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workflow Id"
            }
          },
          {
            "name": "task_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Task Id"
            }
          },
          {
            "name": "object_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Object Id"
            }
          },
          {
            "name": "filter_json",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Filter Json"
            }
          },
          {
            "name": "include_data",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{object_id}": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Get Dataset",
        "description": "Get dataset object metadata by id and optional version.",
        "operationId": "get_dataset_api_v1_datasets__object_id__get",
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Id"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          },
          {
            "name": "include_data",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Dataset Api V1 Datasets  Object Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{object_id}/versions/{version}": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Get Dataset Version",
        "description": "Get a specific dataset object version.",
        "operationId": "get_dataset_version_api_v1_datasets__object_id__versions__version__get",
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Id"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          },
          {
            "name": "include_data",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Data"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Dataset Version Api V1 Datasets  Object Id  Versions  Version  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{object_id}/download": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "Download Dataset",
        "description": "Download dataset payload as a binary attachment.",
        "operationId": "download_dataset_api_v1_datasets__object_id__download_get",
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Id"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/query": {
      "post": {
        "tags": [
          "datasets"
        ],
        "summary": "Query Datasets",
        "description": "Run an advanced read-only query for dataset objects.",
        "operationId": "query_datasets_api_v1_datasets_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObjectQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/query/{scope}": {
      "post": {
        "tags": [
          "query"
        ],
        "summary": "Query Scope",
        "description": "Run a read-only advanced query over a constrained collection scope.",
        "operationId": "query_scope_api_v1_query__scope__post",
        "parameters": [
          {
            "name": "scope",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "workflows",
                "tasks",
                "objects",
                "models",
                "datasets"
              ],
              "type": "string",
              "title": "Scope"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObjectQueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AggregationSpec": {
        "properties": {
          "operator": {
            "type": "string",
            "enum": [
              "avg",
              "sum",
              "min",
              "max"
            ],
            "title": "Operator"
          },
          "field": {
            "type": "string",
            "minLength": 1,
            "title": "Field"
          }
        },
        "type": "object",
        "required": [
          "operator",
          "field"
        ],
        "title": "AggregationSpec",
        "description": "Aggregation operator and source field."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ListResponse": {
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Items"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "count",
          "limit"
        ],
        "title": "ListResponse",
        "description": "Generic list envelope for collection endpoints."
      },
...
      "ObjectQueryRequest": {
        "properties": {
          "filter": {
            "additionalProperties": true,
            "type": "object",
            "title": "Filter"
          },
          "projection": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projection"
          },
          "limit": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Limit",
            "default": 100
          },
          "sort": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SortSpec"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort"
          },
          "aggregation": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AggregationSpec"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aggregation"
          },
          "remove_json_unserializables": {
            "type": "boolean",
            "title": "Remove Json Unserializables",
            "default": true
          },
          "include_data": {
            "type": "boolean",
            "title": "Include Data",
            "default": false
          }
        },
        "type": "object",
        "title": "ObjectQueryRequest",
        "description": "Object query payload with optional payload inclusion."
      },
...
      "SortSpec": {
        "properties": {
          "field": {
            "type": "string",
            "minLength": 1,
            "title": "Field"
          },
          "order": {
            "type": "integer",
            "enum": [
              1,
              -1
            ],
            "title": "Order",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "field"
        ],
        "title": "SortSpec",
        "description": "Sort field/order pair."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}